refactor(observability): split OpenPanel CLI into common/ logic + tools/ wrapper
The tooling gate test_AC8_13_56 requires every tools/*.py to be a thin
wrapper that bootstraps the repo root (ROOT_DIR + sys.path[0]) and re-exports
main(). Move the CLI logic to common/observability/openpanel_query.py
(measured 100% under the common coverage component) and make
tools/openpanel_query.py the standard wrapper. run() now resolves post_json
at call time so it is monkeypatchable in tests.
